src/requirements.txt
portalocker==1.1.0
cloudshell-tg-breaking-point>=1.0,<1.1
cloudshell-shell-core>=3.1,<3.2